% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_S}
\alias{create_S}
\title{Create S}
\usage{
create_S(k_1, k_2, last_trees, num_trees)
}
\arguments{
\item{k_1}{The current value of k1}

\item{k_2}{The current value of k2}

\item{last_trees}{The last trees}

\item{num_trees}{The number of trees}
}
\description{
Creates the S matrix
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
