% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_structures.R
\name{create_stump}
\alias{create_stump}
\alias{update_tree}
\title{Create tree stumps}
\usage{
create_stump(num_trees, num_groups, y, X)

update_tree(
  X,
  y,
  num_groups,
  type = c("grow", "prune", "change", "swap"),
  curr_tree,
  node_min_size
)
}
\arguments{
\item{num_trees}{The number of trees}

\item{num_groups}{The number of groups}

\item{y}{The response variable}

\item{X}{The set of covariates}

\item{type}{The action: grow, prune, change or swap}

\item{curr_tree}{The current state of the tree}

\item{node_min_size}{The minimum number of observations in 
each terminal node}
}
\description{
A function that returns a tree stump (initial tree) given
# the model characteristics

A function that returns a tree stump (initial tree) given
# the model characteristics
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell

Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
