% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{density_plot}
\alias{density_plot}
\title{Density plots for tau and sigma_phi}
\usage{
density_plot(model, type = "tau", sqrt = FALSE)
}
\arguments{
\item{model}{The hebart model object}

\item{type}{Type of plot: tau or sigma_phi}

\item{sqrt}{Logical to decide whether to plot 1/sqrt(tau) instead
(residual precision)}
}
\value{
The correspondent plot
}
\description{
Plots the sampled of values as a density
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
