% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{full_conditional_hebart}
\alias{full_conditional_hebart}
\title{Full conditional of all the data}
\usage{
full_conditional_hebart(y, k_1, k_2, last_trees, num_trees, tau)
}
\arguments{
\item{y}{The y vector}

\item{k_1}{The current value of k_1}

\item{k_2}{The current value of k_2}

\item{last_trees}{The last set of trees}

\item{num_trees}{The number of trees}

\item{tau}{The current value of tau}
}
\description{
A function that returns the full conditional value
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
