% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{get_children}
\alias{get_children}
\title{Get children}
\usage{
get_children(tree_mat, parent)
}
\arguments{
\item{tree_mat}{The tree matrix}

\item{parent}{The corresponding parent}
}
\description{
Gets tree children
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
