% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{HEBART Group predictions}
\usage{
get_predictions(trees, X, single_tree = FALSE)
}
\arguments{
\item{trees}{The current trees}

\item{X}{The set of covariates}

\item{single_tree}{Logical to indicate whether we only have one tree}
}
\description{
A function that returns the group predictions for a hebart model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
