% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{get_tree_prior}
\alias{get_tree_prior}
\title{Get tree prior}
\usage{
get_tree_prior(tree, alpha, beta)
}
\arguments{
\item{tree}{The current tree}

\item{alpha}{The prior value for alpha (grow)}

\item{beta}{The prior value for beta (grow)}
}
\description{
Returns prior distribution value for a tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
