% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_structures.R
\name{grow_tree}
\alias{grow_tree}
\title{Grows a tree}
\usage{
grow_tree(X, y, num_groups, curr_tree, node_min_size)
}
\arguments{
\item{X}{The set of covariates}

\item{y}{The response variable}

\item{num_groups}{The number of groups}

\item{curr_tree}{The current state of the tree}

\item{node_min_size}{The minimum number of observations in 
each terminal node}
}
\description{
A function that returns a grown tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
