% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebart.R
\name{hebart}
\alias{hebart}
\title{Hierarchical Embedded Bayesian Additive Regression Trees}
\usage{
hebart(
  formula,
  data,
  group_variable,
  num_trees = 2,
  control = list(node_min_size = 5),
  priors = list(alpha = 0.95, beta = 2, k_1 = 1, k_2 = 0.2, nu = 3, lambda = 0.1),
  inits = list(tau = 1),
  MCMC = list(iter = 500, burn = 150, thin = 1),
  k_1_pars = list(sample_k1 = TRUE, min_u = 0, max_u = 5, k1_prior = TRUE)
)
}
\arguments{
\item{formula}{The model formula}

\item{data}{The data to be used in the modeling}

\item{group_variable}{The name of the grouping variable}

\item{num_trees}{The number of trees (P)}

\item{control}{A list with control settings}

\item{priors}{A list with prior hyperparameters as defined by the model}

\item{inits}{A list with initial values for parameters}

\item{MCMC}{A list with MCMC parameters}

\item{k_1_pars}{A list to decide whether to sample k1 or not and if yes, from
which range}
}
\value{
A list containing:
 Everything
}
\description{
This function runs a hebart model and returns the tree and
other results obtained in the last iteration of the MCMC
}
\details{
Priors used ----------------------------------------------------------
y_{ij} ~ Normal(m_j, tau^-{1})
tau    ~ Gamma(alpha, beta)
mu     ~ Normal(0, tau_mu = k2*tau^-{1})
mu_j   ~ Normal(mu, k1*tau^-{1})
----------------------------------------------------------------------
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
