% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{logdet}
\alias{logdet}
\title{logdet fuction useful for simple return of log determinant}
\usage{
logdet(A)
}
\arguments{
\item{A}{The matrix}
}
\description{
Gets tree children
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
