% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_hebart}
\alias{predict_hebart}
\title{HEBART Predictions}
\usage{
predict_hebart(
  newX,
  new_groups,
  hebart_posterior,
  type = c("all", "median", "mean")
)
}
\arguments{
\item{newX}{The new set of covariates}

\item{new_groups}{The new groups specification}

\item{hebart_posterior}{The posterior values from the model}

\item{type}{The prediction type ("all", "median" or "mean")}
}
\description{
A function that returns the predictions for a hebart model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
