% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_structures.R
\name{prune_tree}
\alias{prune_tree}
\title{Prunes a tree}
\usage{
prune_tree(X, y, curr_tree)
}
\arguments{
\item{X}{The set of covariates}

\item{y}{The response variable groups}

\item{curr_tree}{The current state of the tree}
}
\description{
A function that returns a pruned tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
