% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{sim_friedman}
\alias{sim_friedman}
\title{Simulate Friedman}
\usage{
sim_friedman(n, pars = c(10, 20, 10, 5), p = 0, scale_err = 1)
}
\arguments{
\item{n}{The number of simulated points}

\item{pars}{The model parameters}

\item{p}{The number of noise variables}

\item{scale_err}{The scale parameter of the simulation}
}
\description{
Simulates data accordingly to the Friedman equation
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
