% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{simulate_mu_groups_hebart}
\alias{simulate_mu_groups_hebart}
\title{Simulate mu_js}
\usage{
simulate_mu_groups_hebart(tree, R, groups, tau, k_1, k_2, num_trees)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{groups}{The groups specification}

\item{tau}{The  current value of tau}

\item{k_1}{The  current value of k_1}

\item{k_2}{The  current value of k_2}

\item{num_trees}{The  number of trees}
}
\description{
Simulates mu_j for each group
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
