% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{simulate_mu_hebart}
\alias{simulate_mu_hebart}
\title{Simulate mu}
\usage{
simulate_mu_hebart(tree, R, tau, k_1, k_2, num_groups)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{tau}{The  current value of tau}

\item{k_1}{The  current value of k_1}

\item{k_2}{The  current value of k_2}

\item{num_groups}{The number of groups}
}
\description{
Simulates mu for each terminal node
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
