% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{simulate_mu_hebart_2}
\alias{simulate_mu_hebart_2}
\title{Simulate mu}
\usage{
simulate_mu_hebart_2(tree, R, tau, k_1, k_2, groups, type, acc)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{tau}{The  current value of tau}

\item{k_1}{The  current value of k_1}

\item{k_2}{The  current value of k_2}

\item{groups}{The groups}

\item{type}{Action type}

\item{acc}{Acceptance or not of proposal}
}
\description{
Simulates mu for each terminal node
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
