% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_structures.R
\name{swap_tree}
\alias{swap_tree}
\title{Swapped a tree}
\usage{
swap_tree(X, y, curr_tree, node_min_size)
}
\arguments{
\item{X}{The set of covariates}

\item{y}{The response variable}

\item{curr_tree}{The current state of the tree}

\item{node_min_size}{The minimum number of observations in
each terminal node}
}
\description{
A function that returns a swapped tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
