% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{traceplot}
\alias{traceplot}
\title{Traceplot for tau or sigma_phi}
\usage{
traceplot(model, type = "tau", sqrt = FALSE)
}
\arguments{
\item{model}{The hebart model object}

\item{type}{Type of plot: tau or sigma_phi}

\item{sqrt}{Logical to decide whether to plot 1/sqrt(tau) instead
(residual precision)}
}
\value{
The correspondent plot
}
\description{
Plots the sampled of values of tau, k1 or sqrt(k1/tau) per iteration
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
