% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{tree_full_conditional}
\alias{tree_full_conditional}
\title{Tree full conditional}
\usage{
tree_full_conditional(tree, R, tau, tau_mu)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{tau}{The current value of tau}

\item{tau_mu}{The current value of tau_mu}
}
\description{
A function that returns current tree full conditional 
distribution value
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
