% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{tree_full_conditional_hebart}
\alias{tree_full_conditional_hebart}
\title{Tree full conditional for hebart}
\usage{
tree_full_conditional_hebart(tree, R, k_1, k_2, M, nu, lambda)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{k_1}{The current value of k_1}

\item{k_2}{The current value of k_2}

\item{M}{The group matrix}

\item{nu}{The current value of nu}

\item{lambda}{The current value of lambda}
}
\description{
A function that returns current tree full conditional
distribution value
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
