% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_k1}
\alias{update_k1}
\title{Update k1}
\usage{
update_k1(y, min_u, max_u, k_1, k_2, last_trees, num_trees, tau, prior = TRUE)
}
\arguments{
\item{y}{The y vector}

\item{min_u}{The lower bound of the sampled values of k1}

\item{max_u}{The upper bound of the sampled values of k1}

\item{k_1}{The current value of k_1}

\item{k_2}{The current value of k_2}

\item{last_trees}{The last set of trees}

\item{num_trees}{The number of trees}

\item{tau}{The current value of tau}

\item{prior}{Logical to decide whether to use a prior (set as
dweibull(1, 5) for now)}
}
\description{
Samples values from the posterior distribution of k1
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
