% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_sigma_phi}
\alias{update_sigma_phi}
\title{Update sigma_phi}
\usage{
update_sigma_phi(
  y,
  S1,
  S2,
  sigma_phi,
  tau_mu,
  tau,
  shape_sigma_phi,
  scale_sigma_phi,
  num_trees,
  sigma_phi_sd = 0.1
)
}
\arguments{
\item{y}{The response vector}

\item{S1}{The group level node allocation matrix}

\item{S2}{The terminal node allocation}

\item{sigma_phi}{The current value of sigma_phi}

\item{tau_mu}{The current value of tau_mu}

\item{tau}{The current value of tau}

\item{shape_sigma_phi}{Weibull shape parameter}

\item{scale_sigma_phi}{Weibull scale parameter}

\item{num_trees}{Number of trees}

\item{sigma_phi_sd}{Standard deviation of proposal distribution}
}
\description{
Samples values from the posterior distribution of sigma_phi
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
