% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_tau}
\alias{update_tau}
\title{Update tau}
\usage{
update_tau(S, nu, lambda, n)
}
\arguments{
\item{S}{The sum of squared residuals}

\item{nu}{.}

\item{lambda}{.}

\item{n}{The number of observations}
}
\description{
Samples values from the posterior distribution of tau
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
