% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_tau_2}
\alias{update_tau_2}
\title{Update tau (different version)}
\usage{
update_tau_2(y, M, nu, lambda, n, groups, k_1, k_2)
}
\arguments{
\item{y}{The response variable}

\item{M}{the binary grouping matrix}

\item{nu}{The current value of nu}

\item{lambda}{The current value of lambda}

\item{n}{The number of observations}

\item{groups}{The grouping variable}

\item{k_1}{The current value of k1}

\item{k_2}{The current value of k2}
}
\description{
Samples values from the posterior distribution of tau
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
