% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_tau_phi}
\alias{update_tau_phi}
\title{Update tau_phi}
\usage{
update_tau_phi(y, tau_phi, last_trees, num_trees, tau, prior = TRUE)
}
\arguments{
\item{y}{The y vector}

\item{last_trees}{The last set of trees}

\item{num_trees}{The number of trees}

\item{tau}{The current value of tau}

\item{prior}{Logical to decide whether to use a prior (set as
dweibull(1, 5) for now)}

\item{min_u}{The lower bound of the sampled values of tau_phi}

\item{max_u}{The upper bound of the sampled values of tau_phi}

\item{k_1}{The current value of k_1}

\item{k_2}{The current value of k_2}
}
\description{
Samples values from the posterior distribution of tau_phi
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
