% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_structures.R
\name{update_tree}
\alias{update_tree}
\title{Create tree stumps}
\usage{
update_tree(
  y,
  X,
  groups,
  type = c("grow", "prune", "change", "swap"),
  curr_tree,
  node_min_size
)
}
\arguments{
\item{y}{The response variable}

\item{X}{The set of covariates}

\item{groups}{The groups}

\item{type}{The action: grow, prune, change or swap}

\item{curr_tree}{The current state of the tree}

\item{node_min_size}{The minimum number of observations in
each terminal node}
}
\description{
A function that returns a tree stump (initial tree) given
# the model characteristics
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
