% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_S}
\alias{create_S}
\title{Create S}
\usage{
create_S(curr_trees, groups = NULL)
}
\arguments{
\item{curr_trees}{The current set of trees}

\item{groups}{if NULL just creates the terminal node matrix. If present 
allocates it into groups too}
}
\description{
Creates the S matrix
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
