% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_structures.R
\name{create_stump}
\alias{create_stump}
\title{Create tree stumps}
\usage{
create_stump(num_trees, groups, y, X)
}
\arguments{
\item{num_trees}{The number of trees}

\item{groups}{The groups}

\item{y}{The response variable}

\item{X}{The set of covariates}
}
\description{
A function that returns a tree stump (initial tree) given
# the model characteristics
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
