% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{full_conditional_hebart}
\alias{full_conditional_hebart}
\title{Tree full conditional for hebart}
\usage{
full_conditional_hebart(tree, R, num_trees, tau, tau_phi, tau_mu, M)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{num_trees}{The number of trees}

\item{tau}{The current value of tau}

\item{tau_phi}{The current value of tau_phi}

\item{tau_mu}{The current value of tau_mu}

\item{M}{The group matrix}
}
\description{
A function that returns current tree full conditional
distribution value
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
