% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{get_group_predictions}
\alias{get_group_predictions}
\title{HEBART Group predictions}
\usage{
get_group_predictions(trees, X, groups, single_tree = FALSE, old_groups)
}
\arguments{
\item{trees}{The current trees}

\item{X}{The set of covariates}

\item{groups}{The groups specification}

\item{single_tree}{Logical to indicate whether we only have one tree}

\item{old_groups}{The groups used during training}
}
\description{
A function that returns the group predictions for a hebart model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
