% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{simulate_mu_hebart}
\alias{simulate_mu_hebart}
\title{Simulate mu}
\usage{
simulate_mu_hebart(tree, R, tau, tau_phi, tau_mu, M, num_trees)
}
\arguments{
\item{tree}{The current tree}

\item{R}{The corresponding residuals for the tree}

\item{tau}{The  current value of tau}

\item{tau_phi}{The current value of tau_phi}

\item{tau_mu}{The value of tau_mu}

\item{M}{The value of M}

\item{num_trees}{The number of trees}
}
\description{
Simulates mu for each terminal node
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
