% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_k1}
\alias{update_k1}
\title{Update k1}
\usage{
update_k1(y, min_u, max_u, k_1, k_2, M, nu, lambda, prior = TRUE)
}
\arguments{
\item{y}{The y vector}

\item{min_u}{The lower bound of the sampled values of k1}

\item{max_u}{The upper bound of the sampled values of k1}

\item{k_1}{The current value of k_1}

\item{k_2}{The current value of k_2}

\item{M}{The group matrix}

\item{nu}{The current value of nu}

\item{lambda}{The current value of lambda}

\item{prior}{Logical to decide whether to use a prior (set as 
dweibull(7, 10) for now)}
}
\description{
Samples values from the posterior distribution of k1
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
