% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_tau}
\alias{update_tau}
\title{Update tau}
\usage{
update_tau(y, M, nu, lambda, num_groups, k_1, k_2, num_trees, last_trees)
}
\arguments{
\item{y}{The y vector}

\item{M}{The current group model matrix}

\item{nu}{The current value of nu}

\item{lambda}{The current value of lambda}

\item{num_groups}{The number of groups}

\item{k_1}{The current value of k1}

\item{k_2}{The current value of k2
# Update tau --------------------------------------------------------------}
}
\description{
Samples values from the posterior distribution of tau
}
\details{
# Simulate mu groups HEBART -----------------------------------------------
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
