% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{update_tau}
\alias{update_tau}
\title{Update tau}
\usage{
update_tau(y, predictions, nu, lambda)
}
\arguments{
\item{y}{The y vector}

\item{predictions}{The current set of predictions}

\item{nu}{The value of nu}

\item{lambda}{The value of lambda}
}
\description{
Samples values from the posterior distribution of tau
}
\details{
# Simulate tau -----------------------------------------------
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}, Andrew Parnell
}
